# Simplify the loop for parsing Integer string