There are several branches in the loop that parse Integer string.
In order to speed up parsing, this patch will move unwanted branches out of the loop.

−       | before | after  | result
--       | --     | --     | --
Oj.load  | 4.926M | 5.164M | 1.048x

### Environment
- Linux
  - Manjaro Linux x86_64
  - Kernel: 5.18.0-1-rt11-MANJARO
  - AMD Ryzen 7 5700G
  - gcc version 12.1.0
  - Ruby 3.1.2

### Before
```
Warming up --------------------------------------
             Oj.load   488.938k i/100ms
Calculating -------------------------------------
             Oj.load      4.926M (± 0.2%) i/s -     49.383M in  10.024672s
```

### After
```
Warming up --------------------------------------
             Oj.load   514.179k i/100ms
Calculating -------------------------------------
             Oj.load      5.164M (± 0.1%) i/s -     51.932M in  10.057128s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'
require 'json'

json = (2 ** 30).to_json

Benchmark.ips do |x|
  x.warmup = 3
  x.time = 10

  x.report('Oj.load') do |times|
    i = 0
    while i < times
      Oj.load(json)
      i += 1
    end
  end
end
```